+++ src/main/java/de/hytalede/event/system/commands/EventCommand.java
@@ -1,0 +1,31 @@
+package de.hytalede.event.system.commands;
+
+import com.hypixel.hytale.component.Ref;
+import com.hypixel.hytale.component.Store;
+import com.hypixel.hytale.protocol.packets.interface_.CustomPageLifetime;
+import com.hypixel.hytale.server.core.command.system.CommandContext;
+import com.hypixel.hytale.server.core.command.system.basecommands.AbstractPlayerCommand;
+import com.hypixel.hytale.server.core.entity.entities.Player;
+import com.hypixel.hytale.server.core.universe.PlayerRef;
+import com.hypixel.hytale.server.core.universe.world.World;
+import com.hypixel.hytale.server.core.universe.world.storage.EntityStore;
+import de.hytalede.event.system.ui.EventUi;
+import org.checkerframework.checker.nullness.compatqual.NonNullDecl;
+
+import java.util.concurrent.CompletableFuture;
+
+public class EventCommand extends AbstractPlayerCommand {
+
+    public EventCommand(String name, String description) {
+        super(name, description);
+    }
+
+    @Override
+    protected void execute(@NonNullDecl CommandContext commandContext, @NonNullDecl Store<EntityStore> store, @NonNullDecl Ref<EntityStore> ref, @NonNullDecl PlayerRef playerRef, @NonNullDecl World world) {
+        Player player = commandContext.senderAs(Player.class);
+
+        CompletableFuture.runAsync(() -> {
+            player.getPageManager().openCustomPage(ref, store, new EventUi(playerRef, CustomPageLifetime.CanDismiss));
+        }, world);
+    }
+}